split state test wasm into a separated file
