+++ dan_layer/engine/tests/state/src/lib.rs
@@ -22,9 +22,12 @@
 
 // TODO: we should only use stdlib if the template dev needs to include it e.g. use core::mem when stdlib is not
 // available
-use std::{mem, ptr::copy, vec::Vec, collections::HashMap};
+use std::{mem, ptr::copy, vec::Vec};
 
-use tari_template_abi::{FunctionDef, Type, TemplateDef, encode_with_len, CallInfo, decode};
+use tari_template_abi::{FunctionDef, Type, encode_with_len, decode};
+use wasm::{generate_abi, TemplateImpl, generate_main};
+
+mod wasm;
 
 // that's what the example should look like from the user's perspective
 #[allow(dead_code)]
@@ -119,57 +122,6 @@ extern "C" fn State_main(call_info: *mut u8, call_info_len: usize) -> *mut u8 {
 }
 
 // TODO: ------ Everything below here should be in a common wasm lib ------
-
-fn generate_abi(template_name: String, functions: Vec<FunctionDef>) -> *mut u8 {
-    let template = TemplateDef {
-        template_name,
-        functions,
-    };
-
-    let buf = encode_with_len(&template);
-    wrap_ptr(buf)
-}
-
-type FunctionImpl = Box<dyn Fn(Vec<Vec<u8>>) -> Vec<u8>>;
-
-struct TemplateImpl(HashMap<String, FunctionImpl>);
-
-impl TemplateImpl {
-    pub fn new() -> Self {
-        Self(HashMap::new())
-    }
-
-    pub fn add_function(&mut self, name: String, implementation: FunctionImpl) {
-        self.0.insert(name.clone(), implementation);
-    }
-}
-
-fn generate_main(call_info: *mut u8, call_info_len: usize, template_impl: TemplateImpl) -> *mut u8 {
-    if call_info.is_null() {
-        panic!("call_info is null");
-    }
-
-    let call_data = unsafe { Vec::from_raw_parts(call_info, call_info_len, call_info_len) };
-    let call_info: CallInfo = decode(&call_data).unwrap();
-
-    // get the function
-    let function = match template_impl.0.get(&call_info.func_name) {
-        Some(f) => f.clone(),
-        None => panic!("invalid function name"),
-    };
-
-    // call the function
-    let result = function(call_info.args);
-
-    // return the encoded results of the function call
-    wrap_ptr(result)
-}
-
-fn wrap_ptr(mut v: Vec<u8>) -> *mut u8 {
-    let ptr = v.as_mut_ptr();
-    mem::forget(v);
-    ptr
-}
 
 extern "C" {
     pub fn tari_engine(op: u32, input_ptr: *const u8, input_len: usize) -> *mut u8;
